+++ app/mobile/lib/view/moderatorHomePage/settleViewHome.dart
@@ -4,6 +4,9 @@ import 'package:mobile_app/view/pollView/tagWidget.dart';
 import 'package:mobile_app/view/pollView/userInformationWidget.dart';
 import 'package:intl/intl.dart';
 
+import '../helpers/dateTime.dart';
+import '../helpers/tag.dart';
+
 class SettleViewHome extends StatelessWidget {
   final String pollId;
   final String userName;
@@ -53,7 +56,7 @@ class SettleViewHome extends StatelessWidget {
               TagListWidget(tags: tags, tagColors: tagColors),
               Row(
                 children: [
-                  DateTimeWidget(dateTime: DateTime.parse(dateTime)),
+                  DateTimeWidget(dateTime: DateTime.parse(dateTime), color: navy),
                 ],
               ),
             ]
@@ -65,57 +68,5 @@ class SettleViewHome extends StatelessWidget {
 }
 
 
-class TagListWidget extends StatelessWidget {
-  final List<String> tags;
-  final List<Color> tagColors;
 
-  const TagListWidget({super.key, required this.tags, required this.tagColors});
 
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: Row(
-        children: [
-          for (int i = 0; i < tags.length; i++)
-            Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 4.0),
-                child: TagWidget(tagText: tags[i], tagColor: tagColors[i])),
-        ],
-      ),
-    );
-  }
-}
-
-class DateTimeWidget extends StatelessWidget {
-  final DateTime dateTime;
-
-  const DateTimeWidget({Key? key, required this.dateTime}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    String formattedDateTime = DateFormat.yMMMMd('en_US').add_jm().format(dateTime);
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 25.0),
-      child: Row(
-        children: [
-          Container(
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(20.0),
-              border: Border.all(color: Colors.pink),
-              color: Colors.pink,
-            ),
-            child: Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Text(
-                formattedDateTime,
-                style: const TextStyle(fontSize: 16.0, color: Colors.white),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
