+++ app/mobile/lib/view/pollViewHomePage/pollViewHomePage.dart
@@ -10,6 +10,9 @@ import 'package:mobile_app/view/pollView/pollView.dart';
 import 'package:mobile_app/services/pollViewHomePageLike.dart';
 import 'package:mobile_app/services/pollViewHomePageVote.dart';
 
+import '../helpers/dateTime.dart';
+import '../helpers/tag.dart';
+
 class PollViewHomePage extends StatefulWidget {
   final String pollId;
   final String userName;
@@ -142,7 +145,7 @@ class _PollViewHomePageState extends State<PollViewHomePage> {
           Row(
             children: [
               LikeCountWidget(likeCount: likeCount),
-              DateTimeWidget(dateTime: widget.dateTime),
+              DateTimeWidget(dateTime: DateTime.parse(widget.dateTime), color: navy),
             ],
           ),
           Padding(
@@ -274,57 +277,6 @@ class LikeCountWidget extends StatelessWidget {
                   style: const TextStyle(fontSize: 16.0, color: whitish)),
             ),
           ),
-        ],
-      ),
-    );
-  }
-}
-
-class DateTimeWidget extends StatelessWidget {
-  final String dateTime;
-
-  const DateTimeWidget({super.key, required this.dateTime});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 25.0),
-      child: Row(
-        children: [
-          Container(
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(20.0),
-              border: Border.all(color: pink),
-              color: pink,
-            ),
-            child: Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Text(dateTime,
-                  style: const TextStyle(fontSize: 16.0, color: whitish)),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class TagListWidget extends StatelessWidget {
-  final List<String> tags;
-  final List<Color> tagColors;
-
-  const TagListWidget({super.key, required this.tags, required this.tagColors});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: Row(
-        children: [
-          for (int i = 0; i < tags.length; i++)
-            Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 4.0),
-                child: TagWidget(tagText: tags[i], tagColor: tagColors[i])),
         ],
       ),
     );
